requirements.txt
dnspython==2.4.2
paho-mqtt==1.6.1
pymongo==4.5.0
python-environ==0.4.54
